% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_dataset_and_model.R
\name{import_dataset_and_model}
\alias{import_dataset_and_model}
\title{Converts a single-cell dataset to an scDissector LDM object}
\usage{
import_dataset_and_model(model_version_name, clustering_data_path, umitab,
  cell_to_cluster, cell_to_sample, min_umis = 250, max_umis = 25000,
  ds_numis = c(200, 500, 1000, 2000), insilico_gating = NULL,
  clustAnnots = NA, ds_list = NULL)
}
\arguments{
\item{model_version_name}{user-defined model version name}

\item{clustering_data_path}{scDissector clustering_data folder. An empty folder could be created by create_clustering_data_dir}

\item{umitab}{UMI matrix. Genes are rows while cells are columns. row names are the gene IDs (usually geneSymbols) and column names are cell IDs.}

\item{cell_to_cluster}{a vector that maps from cells IDs to cluster IDs. The vector names are the cell IDs while clusters IDs are the vector values.}

\item{cell_to_sample}{a vector that maps from cells IDs to sample IDs. The vector names are the cell IDs while sample IDs are the vector values.}

\item{min_umis}{lower UMI-count threshold}

\item{max_umis}{upper UMI-count threshold}

\item{ds_numis}{[optional] vector containing UMI counts to down-sample the dataset to.}

\item{insilico_gating}{[optional] A list containing the in-silico gating policy for the imported cells. Each item in the list is a list containg two parameters: (1) genes- a vector defining a gene list (2) interval - a vector of length 2 defining the lower and upper UMI fraction expressed by the input gene list.}

\item{ds_list}{[optional] user-defined downsapling}

\item{excluded_clusters}{[optional] vector with clusters to exclude. Cells associated with these clusters are not loaded.}

\item{clustAnnots=NA}{[optional] cluster annotations vector. Mapping from cluster IDs to user-defined annotations.}
}
\value{
LDM object
}
\description{
Converts a single-cell dataset to an scDissector LDM object
}
