% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_dataset_and_model.R
\name{load_dataset_and_model}
\alias{load_dataset_and_model}
\title{Load dataset and project onto a predefined model}
\usage{
load_dataset_and_model(model_fn, sample_fns, min_umis = 250,
  model_version_name = "", max_umis = 25000, excluded_clusters = NA,
  ds_numis = NA, genes = NULL, max_ncells_per_sample = NA,
  lightweight = F, cell_list = NULL, max_noise_fraction = 0.2)
}
\arguments{
\item{model_fn}{model file-name}

\item{sample_fns}{vector containing compiled sample file-names}

\item{min_umis}{lower UMI-count threshold}

\item{model_version_name}{model version name}

\item{max_umis}{upper UMI-count threshold}

\item{excluded_clusters}{[optional] vector with clusters to exclude. Cells associated with these clusters are not loaded.}

\item{ds_numis}{[optional] vector containing UMI counts to down-sample the dataset to.}

\item{genes}{[optional ] vector specifying genes which will be included in the analysis. If NULL (default) all genes are included. In cases of inconsistencies between the gene lists included in different samples this vector should provide the desired common gene list.}

\item{max_ncells_per_sample}{[optional ] Integer. limits the number of cells randomly selected and loaded per sample. If NA (default) all cells are loaded}

\item{lightweight}{[optional ] Boolean. If true, unnecessary obectes are not loaded. (F is the default)}

\item{max_noise_fraction}{[optional ] numeric maximum value for the alpha parameter. (0.20 in default)}
}
\value{
LDM object
}
\description{
Load dataset and project onto a predefined model
}
