% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loading_samples_and_running_scDissector.R
\name{load_ldm}
\alias{load_ldm}
\title{preload scRNA sample and project onto a model}
\usage{
load_ldm(clustering_data_dir, model_name, sample_names)
}
\arguments{
\item{clustering_data_dir}{path to the folder that cntains the compiled samples, model versions and metadata}

\item{model_name}{model name as defined in clustering_data_dir/model_versions.csv}

\item{sample_names}{vector containing sample names as defined in clustering_data_dir/samples.csv}
}
\value{
ldm object
}
\description{
preload scRNA sample and project onto a model
}
\examples{
# ldm=load_ldm(clustering_data_dir="path_example/clustering_data_exaple",model_name="model_name_example",sample_names=c("sample1_example","sample2_example","sample3_example"))
# ldm can be stored on disk (optional):
# save(ldm,file="ldm_path_example")
# If saved, in the next time you can load the ldm file instead of regenrating it
# load(file="ldm_path_example")
# run scDissector
# run_scDissector(preloaded_data = ldm,clustering_data_path =clustering_data_dir)
}
