% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_scDissector.R
\name{run_scDissector}
\alias{run_scDissector}
\title{Runs scDissector shiny app}
\usage{
run_scDissector(preloaded_data = NULL, clustering_data_path = NULL)
}
\arguments{
\item{preloaded_data}{LDM object (optional)}

\item{clustering_data_path}{path to clustering data folder (optional)}
}
\description{
Runs scDissector shiny app
}
